# Reuse message buffer for streaming