For https://github.com/grpc/grpc-go/issues/1455