@coocood 2017-12-31
@MakMukhi 
The buffer reused is only enabled in ClientStream, not in ServerStream.
This is different than what we have discussed because I think it's very unlikely a client keeps long-lived connections to thousands of servers.