@coocood 2018-01-26
@MakMukhi 
I'll will post the benchmark result later.

And the failed test is 
```
--- FAIL: TestDialWaitsForServerSettings (5.67s)
	clientconn_test.go:159: Error while dialing. Err: context deadline exceeded
	clientconn_test.go:138: Error while accepting. Err: accept tcp 127.0.0.1:44053: use of closed network connection
```
It seems not related to this PR, and it only failed on a particular Go version.