@coocood 2018-01-26
To focus on the issue I've found, I only post 1M response for now.

The result of after:
```
Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_1048576B-Compressor_false: 
50_Latency: 6.8084 ms 	90_Latency: 10.3493 ms 	99_Latency: 111.5983 ms 	Avg latency: 16.0996 ms 	Count: 2485 	5123585 Bytes/op	478 Allocs/op	
Histogram (unit: ms)
Count: 2485  Min:   3.6  Max: 114.2  Avg: 16.10
------------------------------------------------------------
[        3.564034,         3.564035)     1    0.0%    0.0%  
[        3.564035,         3.564042)     0    0.0%    0.0%  
[        3.564042,         3.564095)     0    0.0%    0.0%  
[        3.564095,         3.564514)     0    0.0%    0.0%  
[        3.564514,         3.567793)     0    0.0%    0.0%  
[        3.567793,         3.593464)     0    0.0%    0.0%  
[        3.593464,         3.794469)     3    0.1%    0.2%  
[        3.794469,         5.368326)   237    9.5%    9.7%  #
[        5.368326,        17.691521)  2020   81.3%   91.0%  ########
[       17.691521,       114.181329)   223    9.0%  100.0%  #
[      114.181329,              inf)     1    0.0%  100.0%  
```

The result of before
```
Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_1048576B-Compressor_false: 
50_Latency: 7.5285 ms 	90_Latency: 11.0990 ms 	99_Latency: 112.7684 ms 	Avg latency: 15.2175 ms 	Count: 2629 	5964215 Bytes/op	449 Allocs/op	
Histogram (unit: ms)
Count: 2629  Min:   3.9  Max: 116.1  Avg: 15.22
------------------------------------------------------------
[        3.883256,         3.883257)     1    0.0%    0.0%  
[        3.883257,         3.883264)     0    0.0%    0.0%  
[        3.883264,         3.883318)     0    0.0%    0.0%  
[        3.883318,         3.883738)     0    0.0%    0.0%  
[        3.883738,         3.887039)     0    0.0%    0.0%  
[        3.887039,         3.912927)     0    0.0%    0.0%  
[        3.912927,         4.115956)     3    0.1%    0.2%  
[        4.115956,         5.708251)   311   11.8%   12.0%  #
[        5.708251,        18.196155)  2118   80.6%   92.5%  ########
[       18.196155,       116.135119)   195    7.4%  100.0%  #
[      116.135119,              inf)     1    0.0%  100.0%  
```

The result is odd, reuse buffer branch allocates more than before.
Then I removed the double length, allocate exactly the same length if it's more than cap, the result is much better:
```
Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_1048576B-Compressor_false: 
50_Latency: 6.6709 ms 	90_Latency: 8.8538 ms 	99_Latency: 109.6082 ms 	Avg latency: 8.3528 ms 	Count: 4790 	5062523 Bytes/op	432 Allocs/op	
Histogram (unit: ms)
Count: 4790  Min:   3.4  Max: 114.9  Avg: 8.35
------------------------------------------------------------
[        3.432767,         3.432768)     1    0.0%    0.0%  
[        3.432768,         3.432775)     0    0.0%    0.0%  
[        3.432775,         3.432828)     0    0.0%    0.0%  
[        3.432828,         3.433248)     0    0.0%    0.0%  
[        3.433248,         3.436538)     0    0.0%    0.0%  
[        3.436538,         3.462321)     1    0.0%    0.0%  
[        3.462321,         3.664369)     4    0.1%    0.1%  
[        3.664369,         5.247720)   378    7.9%    8.0%  #
[        5.247720,        17.655699)  4338   90.6%   98.6%  #########
[       17.655699,       114.891160)    67    1.4%  100.0%  
[      114.891160,              inf)     1    0.0%  100.0%  
```