@coocood 2018-01-26
@MakMukhi 
It seems like the allocation number and average latency not stable on each bench test.

It has nothing to do with double length.

I ran many times, the 50_latency reduced about 9%.


Here is the result of different response size:

After:
```
Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_1024B-Compressor_false:
50_Latency: 159.7260 µs 	90_Latency: 222.1620 µs 	99_Latency: 485.1390 µs 	Avg latency: 172.4190 µs 	Count: 231633 	8518 Bytes/op	55 Allocs/op
Histogram (unit: µs)
Count: 231633  Min:  75.9  Max: 1593.7  Avg: 172.42
------------------------------------------------------------
[     75.897000,      75.898000)       1    0.0%    0.0%
[     75.898000,      75.901862)       0    0.0%    0.0%
[     75.901862,      75.920638)       0    0.0%    0.0%
[     75.920638,      76.011923)       0    0.0%    0.0%
[     76.011923,      76.455742)       0    0.0%    0.0%
[     76.455742,      78.613530)       4    0.0%    0.0%
[     78.613530,      89.104400)     158    0.1%    0.1%
[     89.104400,     140.109594)   60321   26.0%   26.1%  ###
[    140.109594,     388.089960)  167790   72.4%   98.5%  #######
[    388.089960,    1593.737000)    3358    1.4%  100.0%
[   1593.737000,            inf)       1    0.0%  100.0%

Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_65535B-Compressor_false:
50_Latency: 656.3540 µs 	90_Latency: 1211.1790 µs 	99_Latency: 102631.8410 µs 	Avg latency: 3609.9480 µs 	Count: 11080 	375379 Bytes/op83 Allocs/op
Histogram (unit: µs)
Count: 11080  Min: 263.9  Max: 104126.2  Avg: 3609.95
------------------------------------------------------------
[      263.865000,       263.866000)      1    0.0%    0.0%
[      263.866000,       263.872775)      0    0.0%    0.0%
[      263.872775,       263.925455)      0    0.0%    0.0%
[      263.925455,       264.335059)      0    0.0%    0.0%
[      264.335059,       267.519856)      2    0.0%    0.0%
[      267.519856,       292.282632)      1    0.0%    0.0%
[      292.282632,       484.820832)   2172   19.6%   19.6%  ##
[      484.820832,      1981.864555)   8495   76.7%   96.3%  ########
[     1981.864555,     13621.839964)     97    0.9%   97.2%
[    13621.839964,    104126.229000)    311    2.8%  100.0%
[   104126.229000,              inf)      1    0.0%  100.0%

Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_1048576B-Compressor_false:
50_Latency: 6.6519 ms 	90_Latency: 9.0018 ms 	99_Latency: 109.1591 ms 	Avg latency: 8.4781 ms 	Count: 4719 	5153053 Bytes/op	430 Allocs/op
Histogram (unit: ms)
Count: 4719  Min:   3.2  Max: 113.8  Avg: 8.48
------------------------------------------------------------
[        3.162232,         3.162233)     1    0.0%    0.0%
[        3.162233,         3.162240)     0    0.0%    0.0%
[        3.162240,         3.162293)     0    0.0%    0.0%
[        3.162293,         3.162712)     0    0.0%    0.0%
[        3.162712,         3.165991)     0    0.0%    0.0%
[        3.165991,         3.191661)     0    0.0%    0.0%
[        3.191661,         3.392662)     1    0.0%    0.0%
[        3.392662,         4.966478)   192    4.1%    4.1%
[        4.966478,        17.289304)  4451   94.3%   98.4%  #########
[       17.289304,       113.775874)    73    1.5%  100.0%
[      113.775874,              inf)     1    0.0%  100.0%
```

Before:
```
Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_1024B-Compressor_false:
50_Latency: 162.9540 µs 	90_Latency: 233.4670 µs 	99_Latency: 525.4040 µs 	Avg latency: 177.6840 µs 	Count: 224778 	9666 Bytes/op	56 Allocs/op
Histogram (unit: µs)
Count: 224778  Min:  71.2  Max: 2502.2  Avg: 177.68
------------------------------------------------------------
[     71.225000,      71.226000)       1    0.0%    0.0%
[     71.226000,      71.230123)       0    0.0%    0.0%
[     71.230123,      71.251246)       0    0.0%    0.0%
[     71.251246,      71.359461)       1    0.0%    0.0%
[     71.359461,      71.913856)       0    0.0%    0.0%
[     71.913856,      74.754073)       2    0.0%    0.0%
[     74.754073,      89.304767)     189    0.1%    0.1%
[     89.304767,     163.849307)  114519   50.9%   51.0%  #####
[    163.849307,     545.747826)  108026   48.1%   99.1%  #####
[    545.747826,    2502.249000)    2039    0.9%  100.0%
[   2502.249000,            inf)       1    0.0%  100.0%

Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_65535B-Compressor_false:
50_Latency: 777.7100 µs 	90_Latency: 1360.8740 µs 	99_Latency: 102505.5090 µs 	Avg latency: 3696.4090 µs 	Count: 10821 	444362 Bytes/op84 Allocs/op
Histogram (unit: µs)
Count: 10821  Min: 304.8  Max: 105989.1  Avg: 3696.41
------------------------------------------------------------
[      304.817000,       304.818000)      1    0.0%    0.0%
[      304.818000,       304.824790)      0    0.0%    0.0%
[      304.824790,       304.877689)      0    0.0%    0.0%
[      304.877689,       305.289792)      0    0.0%    0.0%
[      305.289792,       308.500213)      0    0.0%    0.0%
[      308.500213,       333.510505)      7    0.1%    0.1%
[      333.510505,       528.349317)   2143   19.8%   19.9%  ##
[      528.349317,      2046.210983)   8222   76.0%   95.9%  ########
[     2046.210983,     13870.877808)    144    1.3%   97.2%
[    13870.877808,    105989.118000)    303    2.8%  100.0%
[   105989.118000,              inf)      1    0.0%  100.0%

Stream-traceMode_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_4-reqSize_64B-respSize_1048576B-Compressor_false:
50_Latency: 7.5659 ms 	90_Latency: 12.1183 ms 	99_Latency: 112.5573 ms 	Avg latency: 13.3949 ms 	Count: 3005 	5975819 Bytes/op	435 Allocs/op
Histogram (unit: ms)
Count: 3005  Min:   3.7  Max: 117.6  Avg: 13.39
------------------------------------------------------------
[        3.671680,         3.671681)     1    0.0%    0.0%
[        3.671681,         3.671688)     0    0.0%    0.0%
[        3.671688,         3.671742)     0    0.0%    0.0%
[        3.671742,         3.672165)     0    0.0%    0.0%
[        3.672165,         3.675489)     0    0.0%    0.0%
[        3.675489,         3.701603)     0    0.0%    0.0%
[        3.701603,         3.906753)     3    0.1%    0.1%
[        3.906753,         5.518406)   198    6.6%    6.7%  #
[        5.518406,        18.179525)  2626   87.4%   94.1%  #########
[       18.179525,       117.645016)   176    5.9%  100.0%  #
[      117.645016,              inf)     1    0.0%  100.0%
```